Fixed VS build
git-svn-id: https://svnsrv.germanlloyd.org/svn/ExtFEMIO/trunk@678 389f8db8-dd28-11e4-b689-8b400cb78fb4

+++ src/bdf/bdf_cards.cpp
@@ -291,7 +291,7 @@ std::ostream &operator<< (std::ostream& os, cards::types cardtype) {
     case cards::types::UNKNOWN: return os << "UNKNOWN";
     // omit default case to trigger compiler warning for missing cases
     };
-    return os << static_cast<std::uint16_t>(cardtype);
+    return os << static_cast<size_t>(cardtype);
 }
